Pushing some import issues with the syntax highlighters.
