+++ HFX/hfx_gui/MultiLineEdit.py
@@ -1,7 +1,7 @@
 # PySide import
-from PySide import QtGui, QtCore
+from PySide import QtGui
 
-from utilities import ConvertToHFX
+from utilities import ConvertToHFX, PythonHighlighter, JSONHighlighter, XMLHighlighter
 
 
 __all__ = [
@@ -29,471 +29,4 @@ class MultiLineEdit(ConvertToHFX, QtGui.QTextEdit):
 
         if syntax:
             syntax(self.document())
-
-
-def format(color, style=''):
-    """Return a QTextCharFormat with the given attributes.
-    """
-    _color = QtGui.QColor()
-    _color.setNamedColor(color)
-
-    _format = QtGui.QTextCharFormat()
-    _format.setForeground(_color)
-    if 'bold' in style:
-        _format.setFontWeight(QtGui.QFont.Bold)
-    if 'italic' in style:
-        _format.setFontItalic(True)
-
-    return _format
-
-
-# Syntax styles that can be shared by all languages
-STYLES = {
-    'keyword': format('#CC6600', 'bold'),
-    'operator': format('red'),
-    'brace': format('darkGray'),
-    'defclass': format('green', 'bold'),
-    'string': format('darkMagenta'),
-    'string2': format('darkMagenta'),
-    'comment': format('darkGreen', 'italic'),
-    'self': format('green', 'italic'),
-    'numbers': format('brown'),
-}
-
-
-class PythonHighlighter(QtGui.QSyntaxHighlighter):
-    """Syntax highlighter for the Python language.
-    """
-    # Python + numpy keywords
-    python_keywords = [
-        'and', 'assert', 'break', 'class', 'continue', 'def',
-        'del', 'elif', 'else', 'except', 'exec', 'finally',
-        'for', 'from', 'global', 'if', 'import', 'in',
-        'is', 'lambda', 'not', 'or', 'pass', 'print',
-        'raise', 'return', 'try', 'while', 'yield',
-        'None', 'True', 'False','linspace','array','arange','zeros']
-
-
-    keywords=python_keywords
-    # Python operators
-    operators = [
-        '=',
-        # Comparison
-        '==', '!=', '<', '<=', '>', '>=',
-        # Arithmetic
-        '\+', '-', '\*', '/', '//', '\%', '\*\*',
-        # In-place
-        '\+=', '-=', '\*=', '/=', '\%=',
-        # Bitwise
-        '\^', '\|', '\&', '\~', '>>', '<<',
-    ]
-
-    # Python braces
-    braces = [
-        '\{', '\}', '\(', '\)', '\[', '\]',
-    ]
-    def __init__(self, document):
-        QtGui.QSyntaxHighlighter.__init__(self, document)
-
-        # Multi-line strings (expression, flag, style)
-        # syntax highlighting from this point onward
-        self.tri_single = (QtCore.QRegExp("'''"), 1, STYLES['string2'])
-        self.tri_double = (QtCore.QRegExp('"""'), 2, STYLES['string2'])
-
-        rules = []
-
-        # Keyword, operator, and brace rules
-        rules += [(r'\b%s\b' % w, 0, STYLES['keyword'])
-            for w in PythonHighlighter.keywords]
-        rules += [(r'%s' % o, 0, STYLES['operator'])
-            for o in PythonHighlighter.operators]
-        rules += [(r'%s' % b, 0, STYLES['brace'])
-            for b in PythonHighlighter.braces]
-
-        # All other rules
-        rules += [
-            # 'self'
-            (r'\bself\b', 0, STYLES['self']),
-
-            # Double-quoted string, possibly containing escape sequences
-            (r'"[^"\\]*(\\.[^"\\]*)*"', 0, STYLES['string']),
-            # Single-quoted string, possibly containing escape sequences
-            (r"'[^'\\]*(\\.[^'\\]*)*'", 0, STYLES['string']),
-
-            # 'def' followed by an identifier
-            (r'\bdef\b\s*(\w+)', 1, STYLES['defclass']),
-            # 'class' followed by an identifier
-            (r'\bclass\b\s*(\w+)', 1, STYLES['defclass']),
-
-            # From '#' until a newline
-            (r'#[^\n]*', 0, STYLES['comment']),
-
-            # Numeric literals
-            (r'\b[+-]?[0-9]+[lL]?\b', 0, STYLES['numbers']),
-            (r'\b[+-]?0[xX][0-9A-Fa-f]+[lL]?\b', 0, STYLES['numbers']),
-            (r'\b[+-]?[0-9]+(?:\.[0-9]+)?(?:[eE][+-]?[0-9]+)?\b', 0, STYLES['numbers']),
-        ]
-
-        # Build a QRegExp for each pattern
-        self.rules = [(QtCore.QRegExp(pat), index, fmt)
-            for (pat, index, fmt) in rules]
-
-
-    def highlightBlock(self, text):
-        """Apply syntax highlighting to the given block of text.
-        """
-        # Do other syntax formatting
-        for expression, nth, format in self.rules:
-            index = expression.indexIn(text, 0)
-
-            while index >= 0:
-                # We actually want the index of the nth match
-                index = expression.pos(nth)
-                length = len(expression.cap(nth))
-                self.setFormat(index, length, format)
-                index = expression.indexIn(text, index + length)
-
-        self.setCurrentBlockState(0)
-
-
-    def match_multiline(self, text, delimiter, in_state, style):
-        """Do highlighting of multi-line strings. ``delimiter`` should be a
-        ``QRegExp`` for triple-single-quotes or triple-double-quotes, and
-        ``in_state`` should be a unique integer to represent the corresponding
-        state changes when inside those strings. Returns True if we're still
-        inside a multi-line string when this function is finished.
-        """
-        # If inside triple-single quotes, start at 0
-        if self.previousBlockState() == in_state:
-            start = 0
-            add = 0
-        # Otherwise, look for the delimiter on this line
-        else:
-            start = delimiter.indexIn(text)
-            # Move past this match
-            add = delimiter.matchedLength()
-
-        # As long as there's a delimiter match on this line...
-        while start >= 0:
-            # Look for the ending delimiter
-            end = delimiter.indexIn(text, start + add)
-            # Ending delimiter on this line?
-            if end >= add:
-                length = end - start + add + delimiter.matchedLength()
-                self.setCurrentBlockState(0)
-            # No; multi-line string
-            else:
-                self.setCurrentBlockState(in_state)
-                try:
-                    length = text.length() - start + add
-                except AttributeError:
-                    pass
-            try:
-                # Apply formatting
-                self.setFormat(start, length, style)
-                # Look for the next match
-                start = delimiter.indexIn(text, start + length)
-            except:
-                pass
-
-        # Return True if still inside a multi-line string, False otherwise
-        if self.currentBlockState() == in_state:
-            return True
-        else:
-            return False
-
-
-def format(color, style=''):
-    """Return a QTextCharFormat with the given attributes.
-    """
-    _color = QtGui.QColor()
-    _color.setNamedColor(color)
-
-    _format = QtGui.QTextCharFormat()
-    _format.setForeground(_color)
-    if 'bold' in style:
-        _format.setFontWeight(QtGui.QFont.Bold)
-    if 'italic' in style:
-        _format.setFontItalic(True)
-
-    return _format
-
-
-# Syntax styles that can be shared by all languages
-STYLES = {
-    'keyword': format('#CC6600', 'bold'),
-    'operator': format('red'),
-    'brace': format('darkGray'),
-    'defclass': format('green', 'bold'),
-    'string': format('darkMagenta'),
-    'string2': format('darkMagenta'),
-    'comment': format('darkGreen', 'italic'),
-    'self': format('green', 'italic'),
-    'numbers': format('brown'),
-}
-
-
-class PythonHighlighter(QtGui.QSyntaxHighlighter):
-    """Syntax highlighter for the Python language.
-    """
-    # Python + numpy keywords
-    python_keywords = [
-        'and', 'assert', 'break', 'class', 'continue', 'def',
-        'del', 'elif', 'else', 'except', 'exec', 'finally',
-        'for', 'from', 'global', 'if', 'import', 'in',
-        'is', 'lambda', 'not', 'or', 'pass', 'print',
-        'raise', 'return', 'try', 'while', 'yield',
-        'None', 'True', 'False','linspace','array','arange','zeros']
-
-
-    keywords=python_keywords
-    # Python operators
-    operators = [
-        '=',
-        # Comparison
-        '==', '!=', '<', '<=', '>', '>=',
-        # Arithmetic
-        '\+', '-', '\*', '/', '//', '\%', '\*\*',
-        # In-place
-        '\+=', '-=', '\*=', '/=', '\%=',
-        # Bitwise
-        '\^', '\|', '\&', '\~', '>>', '<<',
-    ]
-
-    # Python braces
-    braces = [
-        '\{', '\}', '\(', '\)', '\[', '\]',
-    ]
-    def __init__(self, document):
-        QtGui.QSyntaxHighlighter.__init__(self, document)
-
-        # Multi-line strings (expression, flag, style)
-        # syntax highlighting from this point onward
-        self.tri_single = (QtCore.QRegExp("'''"), 1, STYLES['string2'])
-        self.tri_double = (QtCore.QRegExp('"""'), 2, STYLES['string2'])
-
-        rules = []
-
-        # Keyword, operator, and brace rules
-        rules += [(r'\b%s\b' % w, 0, STYLES['keyword'])
-            for w in PythonHighlighter.keywords]
-        rules += [(r'%s' % o, 0, STYLES['operator'])
-            for o in PythonHighlighter.operators]
-        rules += [(r'%s' % b, 0, STYLES['brace'])
-            for b in PythonHighlighter.braces]
-
-        # All other rules
-        rules += [
-            # 'self'
-            (r'\bself\b', 0, STYLES['self']),
-
-            # Double-quoted string, possibly containing escape sequences
-            (r'"[^"\\]*(\\.[^"\\]*)*"', 0, STYLES['string']),
-            # Single-quoted string, possibly containing escape sequences
-            (r"'[^'\\]*(\\.[^'\\]*)*'", 0, STYLES['string']),
-
-            # 'def' followed by an identifier
-            (r'\bdef\b\s*(\w+)', 1, STYLES['defclass']),
-            # 'class' followed by an identifier
-            (r'\bclass\b\s*(\w+)', 1, STYLES['defclass']),
-
-            # From '#' until a newline
-            (r'#[^\n]*', 0, STYLES['comment']),
-
-            # Numeric literals
-            (r'\b[+-]?[0-9]+[lL]?\b', 0, STYLES['numbers']),
-            (r'\b[+-]?0[xX][0-9A-Fa-f]+[lL]?\b', 0, STYLES['numbers']),
-            (r'\b[+-]?[0-9]+(?:\.[0-9]+)?(?:[eE][+-]?[0-9]+)?\b', 0, STYLES['numbers']),
-        ]
-
-        # Build a QRegExp for each pattern
-        self.rules = [(QtCore.QRegExp(pat), index, fmt)
-            for (pat, index, fmt) in rules]
-
-
-    def highlightBlock(self, text):
-        """Apply syntax highlighting to the given block of text.
-        """
-        # Do other syntax formatting
-        for expression, nth, format in self.rules:
-            index = expression.indexIn(text, 0)
-
-            while index >= 0:
-                # We actually want the index of the nth match
-                index = expression.pos(nth)
-                length = len(expression.cap(nth))
-                self.setFormat(index, length, format)
-                index = expression.indexIn(text, index + length)
-
-        self.setCurrentBlockState(0)
-
-
-    def match_multiline(self, text, delimiter, in_state, style):
-        """Do highlighting of multi-line strings. ``delimiter`` should be a
-        ``QRegExp`` for triple-single-quotes or triple-double-quotes, and
-        ``in_state`` should be a unique integer to represent the corresponding
-        state changes when inside those strings. Returns True if we're still
-        inside a multi-line string when this function is finished.
-        """
-        # If inside triple-single quotes, start at 0
-        if self.previousBlockState() == in_state:
-            start = 0
-            add = 0
-        # Otherwise, look for the delimiter on this line
-        else:
-            start = delimiter.indexIn(text)
-            # Move past this match
-            add = delimiter.matchedLength()
-
-        # As long as there's a delimiter match on this line...
-        while start >= 0:
-            # Look for the ending delimiter
-            end = delimiter.indexIn(text, start + add)
-            # Ending delimiter on this line?
-            if end >= add:
-                length = end - start + add + delimiter.matchedLength()
-                self.setCurrentBlockState(0)
-            # No; multi-line string
-            else:
-                self.setCurrentBlockState(in_state)
-                try:
-                    length = text.length() - start + add
-                except AttributeError:
-                    pass
-            try:
-                # Apply formatting
-                self.setFormat(start, length, style)
-                # Look for the next match
-                start = delimiter.indexIn(text, start + length)
-            except:
-                pass
-
-        # Return True if still inside a multi-line string, False otherwise
-        if self.currentBlockState() == in_state:
-            return True
-        else:
-            return False
-
-
-class XMLHighlighter(QtGui.QSyntaxHighlighter):
-    #INIT THE STUFF
-    def __init__(self, parent=None):
-        super(XMLHighlighter, self).__init__(parent)
-
-        keywordFormat = QtGui.QTextCharFormat()
-        keywordFormat.setForeground(QtCore.Qt.darkMagenta)
-        keywordFormat.setFontWeight(QtGui.QFont.Bold)
-
-        keywordPatterns = ["\\b?xml\\b", "/>", ">", "<"]
-
-        self.highlightingRules = [(QtCore.QRegExp(pattern), keywordFormat) for pattern in keywordPatterns]
-
-        xmlElementFormat = QtGui.QTextCharFormat()
-        xmlElementFormat.setFontWeight(QtGui.QFont.Bold)
-        xmlElementFormat.setForeground(QtCore.Qt.green)
-        self.highlightingRules.append((QtCore.QRegExp("\\b[A-Za-z0-9_]+(?=[\s/>])"), xmlElementFormat))
-
-        xmlAttributeFormat = QtGui.QTextCharFormat()
-        xmlAttributeFormat.setFontItalic(True)
-        xmlAttributeFormat.setForeground(QtCore.Qt.blue)
-        self.highlightingRules.append((QtCore.QRegExp("\\b[A-Za-z0-9_]+(?=\\=)"), xmlAttributeFormat))
-
-        self.valueFormat = QtGui.QTextCharFormat()
-        self.valueFormat.setForeground(QtCore.Qt.red)
-
-        self.valueStartExpression = QtCore.QRegExp("\"")
-        self.valueEndExpression = QtCore.QRegExp("\"(?=[\s></])")
-
-        singleLineCommentFormat = QtGui.QTextCharFormat()
-        singleLineCommentFormat.setForeground(QtCore.Qt.gray)
-        self.highlightingRules.append((QtCore.QRegExp("<!--[^\n]*-->"), singleLineCommentFormat))
-
-    #VIRTUAL FUNCTION WE OVERRIDE THAT DOES ALL THE COLLORING
-    def highlightBlock(self, text):
-
-        #for every pattern
-        for pattern, format in self.highlightingRules:
-
-            #Create a regular expression from the retrieved pattern
-            expression = QtCore.QRegExp(pattern)
-
-            #Check what index that expression occurs at with the ENTIRE text
-            index = expression.indexIn(text)
-
-            #While the index is greater than 0
-            while index >= 0:
-
-                #Get the length of how long the expression is true, set the format from the start to the length with the text format
-                length = expression.matchedLength()
-                self.setFormat(index, length, format)
-
-                #Set index to where the expression ends in the text
-                index = expression.indexIn(text, index + length)
-
-        #HANDLE QUOTATION MARKS NOW.. WE WANT TO START WITH " AND END WITH ".. A THIRD " SHOULD NOT CAUSE THE WORDS INBETWEEN SECOND AND THIRD TO BE COLORED
-        self.setCurrentBlockState(0)
-
-        startIndex = 0
-        if self.previousBlockState() != 1:
-            startIndex = self.valueStartExpression.indexIn(text)
-
-        while startIndex >= 0:
-            endIndex = self.valueEndExpression.indexIn(text, startIndex)
-
-            if endIndex == -1:
-                self.setCurrentBlockState(1)
-                commentLength = len(text) - startIndex
-            else:
-                commentLength = endIndex - startIndex + self.valueEndExpression.matchedLength()
-
-            self.setFormat(startIndex, commentLength, self.valueFormat)
-
-            startIndex = self.valueStartExpression.indexIn(text, startIndex + commentLength)
-
-
-class JSONHighlighter(QtGui.QSyntaxHighlighter):
-    def __init__(self, parent=None):
-        """ Constructor
-        """
-        super(JSONHighlighter, self).__init__(parent)
-
-        self.symbol_format = QtGui.QTextCharFormat()
-        self.symbol_format.setForeground(QtCore.Qt.red)
-        self.symbol_format.setFontWeight(QtGui.QFont.Bold)
-
-        self.name_format = QtGui.QTextCharFormat()
-        self.name_format.setForeground(QtCore.Qt.blue)
-        self.name_format.setFontWeight(QtGui.QFont.Bold)
-        self.name_format.setFontItalic(True)
-
-        self.value_format = QtGui.QTextCharFormat()
-        self.value_format.setForeground(QtCore.Qt.darkGreen)
-
-    def highlightBlock(self, text):
-        """ Highlight a block of code using the rules outlined in the Constructor
-        """
-        expression = QtCore.QRegExp("(\\{|\\}|\\[|\\]|\\:|\\,)")
-        index = expression.indexIn(text)
-        while index >= 0:
-            length = expression.matchedLength()
-            self.setFormat(index, length, self.symbol_format)
-            index = expression.indexIn(text, index + length)
-
-        text.replace("\\\"", "  ")
-
-        expression = QtCore.QRegExp("\".*\" *\\:")
-        expression.setMinimal(True)
-        index = expression.indexIn(text)
-        while index >= 0:
-            length = expression.matchedLength()
-            self.setFormat(index, length - 1, self.name_format)
-            index =expression.indexIn(text, index + length)
-
-        expression = QtCore.QRegExp("\\: *\".*\"")
-        expression.setMinimal(True)
-        index = expression.indexIn(text)
-        while index >= 0:
-            length = expression.matchedLength()
-            self.setFormat(index, length - 1, self.value_format)
-            index = expression.indexIn(text, index + length)
