+++ HFX/hfx_gui/utilities/__init__.py
@@ -1,2 +1,3 @@
 from instance import *
 from PySide_to_HFX import ConvertToHFX, applyHFXStyle, Vertical, Horizontal, guiKitVersion
+from highlighters import PythonHighlighter, JSONHighlighter, XMLHighlighter
